Implemented bug fixes on IP Address Detection
When scraping the metadata on ec2 instances, I had to use the WebClient
method to return the IP Address if the script runs as system. I also
made it possible for the script to throw an error if the function failed
to grab an IP Address. ReadMe has been updated accordingly.